+++ qoqo/build.rs
@@ -480,12 +480,11 @@ fn main() {
                                         _ => None
                                     };
                                 }},
-                                "SpinHamiltonian" => {quote!{
+                                "PauliHamiltonian" => {quote!{
                                     let #pyobject_name = &op
                                     .call_method0(#ident_string)
                                     .map_err(|_| QoqoError::ConversionError)?;
-                                    let temp_op: struqture::spins::SpinHamiltonianSystem = struqture_py::spins::SpinHamiltonianSystemWrapper::from_pyany(#pyobject_name).map_err(|_| QoqoError::ConversionError)?;
-                                    let #ident = temp_op.hamiltonian().clone();
+                                    let #ident: struqture::spins::PauliHamiltonian = struqture_py::spins::PauliHamiltonianWrapper::from_pyany(#pyobject_name).map_err(|_| QoqoError::ConversionError)?;
                                 }},
                                 _ => {
                                     quote!{